globally locked Window's redraw... OpenGl is just not really thread safe...
In particular, drawing text to a mapped buffer caused problems when we had
multiple windows with an own thread each. Now, the draw is locked, but the
pipeline updates are still processed by their respective threads in parallel.

It should still be possible to prepare whatever you need for redrawing during
the pipeline update in parallel (e.g., recording display lists).